Move capacity url to a seperate file
