+++ packages/playground/src/components/connect_wallet_landing.vue
@@ -90,24 +90,15 @@
 </template>
 
 <script lang="ts">
-import { NetworkEnv } from "@threefold/grid_client";
 import { useTheme } from "vuetify";
+
+import capacityURL from "@/utils/getCapacityUrl";
 
 export default {
   name: "ConnectWalletLanding",
   setup() {
     const theme = useTheme();
     const baseUrl = import.meta.env.BASE_URL;
-    const network = process.env.NETWORK || (window as any).env.NETWORK;
-
-    const getCapacityURL = (network: NetworkEnv) => {
-      if (network === NetworkEnv.main) {
-        return "https://dashboard.grid.tf/explorer/statistics";
-      }
-      return `https://dashboard.${network}.grid.tf/explorer/statistics`;
-    };
-
-    const capacityURL = getCapacityURL(network);
 
     return {
       theme,
